[feat] systray linux: unlink temp files as soon as they are in use
Stop generating one file in /tmp for each icon change. Let's move the
clean up to the set icon instead of waiting for the quit clean up.
Unlink will remove the file from /tmp, but app_indicator will hold a
descriptor until it finishes with it.

This is a cherry-pick from a pull-request upstream:
https://github.com/getlantern/systray/pull/63

- Resolves: #60