[bug] launch the systray ASAP so notifications work
The linux version of the notifications library we are using doesn't
handle the gtk.main loop. It requires the systray to be running to be
able to display a notification.

Spliting the start of the systray and the loop we can start the systray
pretty early and later on launch the main loop once we have bitmask and
other stuff ready.

- Related: #88