[pkg] customize windows installer
I kind of liked the "classic" style more (looks cleaner), but it looks
like we only can customize the logo etc with the "modern" look and feel
in windows.

I think the way to go would be to add a QStyleSheet to the installer
(and adding logo etc there), but I think that is going to take some
reading. this is not too bad for now, at least we can put logos to
differentiate riseup/calyx etc.